added api

+++ backend-M3/src/main/java/com/fincore/BankingManagement/Beneficiary/Repository/BeneficiaryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BeneficiaryRepository extends JpaRepository<beneficiary,Integer> {
+public interface BeneficiaryRepository extends JpaRepository<beneficiary,Long> {
 }
